Arkon/client/requirements.txt
python-socketio[client]
pyttsx3
pyautogui
requests
